% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_covid_obj.R
\name{write_corona_obj}
\alias{write_corona_obj}
\title{Write Customized SARS COVID-19 3D Model}
\usage{
write_corona_obj(
  filename = "covidcustom.obj",
  color_membrane = "#a1cdf0",
  color_spike = "#3432cf",
  color_open_spike = "#d92bc5"
)
}
\arguments{
\item{filename}{Default `covidcustom.obj`.}

\item{color_membrane}{Default `#a1cdf0`. Color of the main membrane (lipid layer) of the coronavirus.}

\item{color_spike}{Default `#3432cf`. Color of the closed spike protein (crown) on the coronavirus.}

\item{color_open_spike}{Default `#d92bc5`. Color of the open spike protein (crown) on the coronavirus.}
}
\value{
Path to the file.
}
\description{
This function writes out a custom 3D model of the COVID-19 virus, with
colors specified by the user. The user can change the color of the main
body, the color of the spike protein, and the color of an open spike protein.
}
\details{
The base model was sourced from the NIH 3D Print Exchange.
}
\examples{
library(rayrender)
#Write a base version of the model and render it with rayrender, using vertex colors.
\donttest{

write_corona_obj("defaults.obj")

#Render the model:
obj_model("defaults.obj", vertex_colors = TRUE) \%>\%
  add_object(sphere(y=10,z=10,x=10, material=light(color="lightblue",intensity=100))) \%>\%
  add_object(sphere(y=10,z=10,x=-10, material=light(color="orange",intensity=100))) \%>\%
  render_scene(parallel=TRUE, samples = 100, fov = 7, min_variance=0, focal_distance = 9.6,
               width=800,height=800)

#Now write a custom version, recreating the "standard" grey and red COVID-19 visualization.

write_corona_obj("standard.obj",
                color_membrane = "#a1cdf0",
                color_spike = "#3432cf",
                color_open_spike = "#d92bc5")

#Render the custom model:
obj_model("standard.obj", vertex_colors = TRUE) \%>\%
  add_object(sphere(y=10,z=10,x=10, material=light(color="lightblue",intensity=100))) \%>\%
  add_object(sphere(y=10,z=10,x=-10, material=light(color="orange",intensity=100))) \%>\%
  render_scene(parallel=TRUE, samples = 100, fov = 7, min_variance=0, focal_distance = 9.6,
               width=800,height=800)
}
}
